Retrieve list of todos at v1/todos
