+++ app.js
@@ -2,21 +2,13 @@ const express = require('express');
 const app = express();
 const port = 3000;
 
-var todos = [
-    { id: 1, value: "Walk dog" },
-    { id: 2, value: "Buy milk" },
-    { id: 3, value: "Take out trash" },
-]
+var todos = require('./mockup/todos')
 
-var JSONAPISerializer = require('jsonapi-serializer').Serializer;
-
-var TodoSerializer = new JSONAPISerializer('todos', {
-  attributes: ['value']
+app.get('/v1/status', (_, res) => res.status(204).send(''));
+app.get('/v1/todos', function (_, res) {
+    res.header('Content-Type', 'application/vnd.api+json');
+    res.send(todos);
 });
-
-var todos = TodoSerializer.serialize(todos);
-
-app.get('/v1/status', (req, res) => res.status(204).send(''));
 
 app.listen(port, () => console.log(`Listening on port ${port}`));
